소켓 이벤트 처리 로직을 socketHandler로 분리 및 개선
